+++ packages/react-native-sdk/src/components/Call/CallLayout/CallParticipantsSpotlight.tsx
@@ -91,6 +91,11 @@ export const CallParticipantsSpotlight = ({
     marginHorizontal: landscape ? 0 : 8,
   };
 
+  const showShareScreenOverlay =
+    participantInSpotlight?.isLocalParticipant &&
+    isScreenShareOnSpotlight &&
+    ScreenShareOverlay;
+
   return (
     <View
       testID={ComponentTestIds.CALL_PARTICIPANTS_SPOTLIGHT}
@@ -105,7 +110,7 @@ export const CallParticipantsSpotlight = ({
     >
       {participantInSpotlight &&
         ParticipantView &&
-        (participantInSpotlight.isLocalParticipant && ScreenShareOverlay ? (
+        (showShareScreenOverlay ? (
           <ScreenShareOverlay />
         ) : (
           <ParticipantView